+++ NEWS.md
@@ -12,3 +12,8 @@
 ## Version: 1.2.3
 
 * Set VignetteIndexEntry
+
+## Version: 1.2.4
+
+* Giving reports and statistics
+* Iteration 1: if `small.positive` is greater than the smallest positive eigenvalue (x), replace it with x/10 and report the user.